+++ src/GraphQL.php
@@ -11,6 +11,8 @@ use GraphQL\Server\RequestError;
 use GraphQL\Type\Schema;
 use Illuminate\Contracts\Events\Dispatcher as EventDispatcher;
 use Illuminate\Http\Request;
+use Illuminate\Pipeline\Pipeline;
+use Illuminate\Support\Collection;
 use Laragraph\LaravelGraphQLUtils\RequestParser;
 use Nuwave\Lighthouse\Events\BuildExtensionsResponse;
 use Nuwave\Lighthouse\Events\ManipulateResult;
@@ -22,7 +24,6 @@ use Nuwave\Lighthouse\Schema\SchemaBuilder;
 use Nuwave\Lighthouse\Support\Contracts\CreatesContext;
 use Nuwave\Lighthouse\Support\Contracts\GraphQLContext;
 use Nuwave\Lighthouse\Support\Contracts\ProvidesValidationRules;
-use Nuwave\Lighthouse\Support\Pipeline;
 use Nuwave\Lighthouse\Support\Utils as LighthouseUtils;
 
 class GraphQL
@@ -38,7 +39,7 @@ class GraphQL
     protected $schemaBuilder;
 
     /**
-     * @var \Nuwave\Lighthouse\Support\Pipeline
+     * @var \Illuminate\Pipeline\Pipeline
      */
     protected $pipeline;
 
@@ -211,19 +212,26 @@ class GraphQL
             function (array $errors, callable $formatter): array {
                 // User defined error handlers, implementing \Nuwave\Lighthouse\Execution\ErrorHandler
                 // This allows the user to register multiple handlers and pipe the errors through.
-                $handlers = config('lighthouse.error_handlers', []);
+                $handlers = [];
+                foreach (config('lighthouse.error_handlers', []) as $handlerClass) {
+                    $handlers [] = app($handlerClass);
+                }
 
-                return array_map(
-                    function (Error $error) use ($handlers, $formatter) {
+                return (new Collection($errors))
+                    ->map(function (Error $error) use ($handlers, $formatter): ?array {
                         return $this->pipeline
                             ->send($error)
                             ->through($handlers)
-                            ->then(function (Error $error) use ($formatter) {
+                            ->then(function (?Error $error) use ($formatter): ?array {
+                                if ($error === null) {
+                                    return null;
+                                }
+
                                 return $formatter($error);
                             });
-                    },
-                    $errors
-                );
+                    })
+                    ->filter()
+                    ->all();
             }
         );